MODELS/CapsNetBERT/requirements.txt
pytorch-pretrained-bert==0.6.2
adabound==0.0.5
pyyaml==5.1.2
numpy==1.17.2
scikit-learn==0.21.3
scipy==1.3.1
torch==1.1.0
spacy==2.1.8